release/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["skip"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
